training_repo/requirements.txt
python-dotenv
telethon
pinecone-client
google-generativeai